Add NavBackStackSerializer class
Introduces `NavBackStackSerializer`, a `KSerializer` for the
`NavBackStack` class.

This allows `NavBackStack` (which is based on `SnapshotStateList`)
to be serialized using `kotlinx.serialization`. It delegates
element serialization to a provided `KSerializer<T>` and uses
`SnapshotStateListSerializer` internally to handle the list.

RelNote: "Add `NavBackStackSerializer` to allow serializing and
deserializing a `NavBackStack` with `kotlinx.serialization`.
This is necessary for saving and restoring navigation state.
Pass a serializer for your `NavKey` elements to its
constructor."

Bug: 420443609
Test: N/A
Change-Id: I76afca1d05de99e393ff43e3282c5b9e06514294