Refactor `NavigationEventHandler` creation
Move the `ComposeNavigationEventHandler` creation from
`rememberNavigationEventState` into the `NavigationEventHandler`
composable.

This change ties the handler's lifecycle to the composable that
registers it with the dispatcher, rather than to the
`NavigationEventState` object. The `NavigationEventState` now
holds a nullable reference to the handler, which is set in the
`NavigationEventHandler`'s `DisposableEffect`.

For this to work, we also updated the public API variance for
`NavigationEventState` parameters (from `State<*>` to
`State<out NavigationEventInfo>`).

Additionally: moved the `ComposeNavigationEventHandler`
implementation to be a private class within
`NavigationEventHandler.kt`.

RelNote: N/A

Bug: 444734264
Test: NavigationEventHandlerTest.kt
Change-Id: I5c3e25f455f9cdd8f5438a170d08d4dae8538a39