Apply @Direction annotation to parameters, not types
The @IntDef annotation (used by @Direction) does not
function correctly when applied as a type annotation
(@Target(AnnotationTarget.TYPE)).

This change updates the @Direction annotation target to
VALUE_PARAMETER and PROPERTY and moves all usages from
the type to the parameter itself.

This also adds a TRANSITIONING_UNKNOWN state to the
IntDef. This replaces the previous magic '0' default
and avoids using a nullable Int? for the in-progress
direction, preventing unnecessary boxing.

Fixes: 445155869
Test: existing passes
Change-Id: Iaf754847a18074be4643eeff99d5b829df53825e