Use RemoteColor's getIdForCreationState in tween.
The RemoteColor tween function was using `idProvider()` on the `from` and `to` colors instead of their `getIdForCreationState(creationState)` function. This meant that if a color had a non-constant value (i.e. was backed by a lambda that creates a color in the `creationState`), the `idProvider()` would be called multiple times.

This change fixes that by calling `getIdForCreationState` for the `from` and `to` colors. This ensures that the color is created in the `creationState` if it hasn't been already, and then returns the correct ID.

A new test `tweenTwice` is added to verify that when a non-constant color is used multiple times in tweens, its creation lambda is only invoked once.

Test: tweenTwice
Change-Id: Ida7ccce0177a34df10a8bb886426cc63964abd64