Fix race condition in LifecycleCameraProviderImpl.initAsync
There was a race condition in LifecycleCameraProviderImpl.initAsync
where the initFuture returned to the caller could complete before
the onSuccess callback (which calls initInternal()) was executed.

This could lead to a NullPointerException if the caller immediately
used the provider (e.g., by calling hasCamera()), as the internal
cameraX instance would still be null.

This change fixes the race by moving the initInternal() call into
the main ListenableFuture chain using .transform(). This guarantees
that the cameraX instance is set *before* the future returned to the
caller is marked as complete.

Bug: 446803522
Test: blaze --blazerc=/dev/null test //javatests/com/google/android/libraries/video/camera:CameraControllerTest --compilation_mode=opt --android_platforms=//buildenv/platforms/android:x86 --nocache_test_results  --runs_per_test=10000 --runs_per_test_detects_flakes
Change-Id: I7a2a45ce63e24a47be3bb2123012a4f1d55b430c