Removing the term "Mode" from FullSpaceMode and HomeSpaceMode in SpaceModeSwitch. As well as setting the terms "Full Space" and "Home Space" as notranslate in String.xml
The public documentation has removed the word "Mode" when defining the two states. Localization teams were confused how to translate the two terms, so after conversations with UX they have defined the two terms are not to be localized.

Relnote: Removing the term 'Mode' from SpaceToggleButton as well as removing the terms 'Full Space' and 'Home Space' from string localization.

Test: Log statements with both stringResources working correctly
Bug:447216407
Change-Id: I1c0fc8491bf11e357ab8767af2fe2a9d52be29f7