Mutex internal raw connection usages
When Room internally (in codegen) uses the RawConnectionAccessor to access a pool connection it should still use the mutex to avoid
concurrent usages of different statements from the same connection, as otherwise that would violate the multi-thread rules of SQLite and will lead to obscure errors.

Using the mutex for raw connection access exposed a flaw in blocking @Transaction DAO functions with bodies, as those where using performBlocking() and acquiring a raw connection without actually using it. This change also updates those functions to use a new performInTransactionBlocking() that starts a transaction but does not actually hold a raw connection lock.

Bug: 447171815
Test: CoroutineTest + Existing
Change-Id: I876acd4cbf493d8d6bea26bcb7ef0fc969ef0ab0