Adaptively scale the recommended content box to support new and existing devices.
The recommended content box logic previously used a value directly that was only correct on Moohan due to a bug in the Extensions library. If the known "buggy" value is detected, it's used as-is for backward compatibility. For all other values, the size is correctly scaled, preventing incorrect layouts on newer devices (like Aura) and ensuring forward compatibility once the Extensions library is fixed.

When the Extensions buggy box values for Moohan is fixed, the follow work for removing the conditional unscaling is required.

Bug: 440467693
Test: Passing the existing unit tests in Subspace regarding the sizing and ran a manual testapp that fillMaxs out the Subspace()'s default constraints with both buggy values and correct values. Both cases fills up the entire FOV.
Change-Id: I97cbd9f95e391fe9111d8183a077c79732156a9d