Add isLogExcluded to CallAttributesCompat
Introduces an optional `isLogExcluded` Boolean property to the
CallAttributesCompat data class. This allows applications to request
that a call not be persisted in the system call log.

This new property is only effective on devices running
Build.VERSION.SDK_INT_FULL >= Build.VERSION_CODES_FULL.BAKLAVA_1

If `isLogExcluded` is null or the device is on an older platform
version, this setting has no effect on the platform CallAttributes.

Relnote: "Added `CallAttributesCompat.isLogExcluded` to allow specifying whether
a call should be excluded from the call log"
Test: new unit tests in CallAttributesCompatTest
Fixes: 447166999

Change-Id: Ia2a3a86b05f70c4a4313a84031607cbee356011d