Make LocalEntriesToRenderInCurrentScene internal
LocalEntriesToRenderInCurrentScene is a
composition local specifically used between
SceneSetupNavEntryDecorator (which is internal)
and NavDisplay as an optimization to avoid issues
specific to NavDisplay's use of AnimatedContent.

Ideally, this would be moved entirely within
SceneState (which is the public API that internally
uses SceneSetupNavEntryDecorator), but this API
is at best confusing, so removing it from the public
API and making it internal will give us the flexibility
to decouple these two components later.

Relnote: "Custom displays using `SceneState` as an
alternative to using `NavDisplay` no longer are required
to use the `LocalEntriesToRenderInCurrentScene` composition
local, which is now internal."
Test: ./gradlew checkApi
Fixes: 414668196

Change-Id: Ic40ef603fd633af7aca1a4f3ec945f12764b09cf