Remove navEntryDecorator function
Simply the API surface by removing the navEntryDecorator function and make the NavEntryDecorator an open class with public constructor.

Added kdocs on NavEntryDecorator's onPop parameter to clarify the conditions that triggers this lambda. Also added kdocs on NavEntry.contentKey to clarify that decorator states are keyed to contentKey rather than NavEntry instances since entries can share contentKey if they want to share states.

Bug: 444447434
Bug: 447381176
Test: existing tests pass
Relnote: "The navEntryDecorator function which creates and returns a NavEntryDecorator has been removed and replaced by the NavEntryDecorator class which is now made public and open for subclassing."
Change-Id: If81f85846a52e8f70708e51932b4ff132ec6222b