Refactor SavedStateNavEntryDecorator
Renamed it to SaveableStateHolderNavEntryDecorator which more accurately reflects that fact that the entries are decorated with a SaveableStateHolder.

Then refactored the function into a class since it's functionally a factory for NavEntryDecorator.

Also refactored the internal SceneSetupNavEntryDecorator to align with the new NavEntryDecorator implementation pattern.

Bug: 447381176
Test: existing tests pass
Relnote: "SavedStateNavEntryDecorator has been renamed to SaveableStateHolderNavEntryDecorator since it decorates entries with a SaveableStateHolder. The decorator has also been refactored from a function to a class since it is functionally a factory for NavEntryDecorator."
Change-Id: Ie60132e292e0d8ffb4c0c7b5ac219c37822e2028