Fix: Saving after an undo operation incorrectly persists the undone change
When annotation in undone using undo and document is saved, the undone annotation is incorrectly persisted to the file. As remove operation would fail silently caused by a caught IllegalArgumentException from Aosp as Get operation is required before calling remove operation on an re-opened page.

This change adds a defensive call to page.getPageAnnotations() before
the remove operation to ensure the aosp mapping is populated, allowing the
undo to complete successfully.

Test: build and run on local device
Bug: 448064449
Change-Id: If7bcccb31c02f34d0be5ddbf9861e803140a5879