Clean up remaining entity listeners during dispose()
Cleanup ActivitySpace and AnchorEntity listeners.

SpatialEnvironment was already cleaning up properly in the runtime
layer so just updated the kDoc to mention that.

Did not modify any of the components because their listeners can be
reused if they're attached to a new entity.

Bug: 447652647
Test: Ran testapp
Relnote: "All remaining listeners are automatically removed when
 entities are disposed. It's no longer necessary to manually remove
 these listeners during an Activity's OnDestroy()"
Change-Id: I8f8d68251ede935fdad87e09919b2e50233534fc