Add new test-only CameraState listening API
CameraX currently exposes CameraState through LiveData which has a
dependency on main thread on updating values. This can be problematic in
certain testing environments, such as Robolectric tests that start on
the main thread, as it becomes difficult to await camera state updates
without blocking.

This change introduces a new test-only listener-based API on
`CameraInfo` for observing `CameraState` changes. This provides a more
flexible way to monitor camera state in tests, avoiding the main thread
dependency of `LiveData`.

The new API consists of `addCameraStateListener` and
`removeCameraStateListener`, which are annotated with
`@VisibleForTesting` and, for now, `@RestrictTo(Scope.LIBRARY_GROUP)`.

Bug: 446844264
Test: CameraStateTest, CameraStateAdapterTest
Change-Id: I4bad74244ecf44dd97f25af5a5f51e84b57b8b9f