Enable PausableComposition in lazy layouts prefetch by default
This feature was disabled earlier due to discovered bugs, which are all fixed now.

Test: existing tests, tested by our partner teams
Relnote: ComposeFoundationFlags.isPausableCompositionInPrefetchEnabled is enabled by default. It is a performance optimization, which allows us to distribute work we need to do during the prefetch better, for example we can only perform the composition for parts of the LazyColumn's next item during one ui frame, and then continue composing the rest of it in the next frames.
Change-Id: I4c7feb6cff8b812e75cbf0a1d4e512b3c42ee412