Make FakeRuntimeAnchor implement ExportableAnchor.
Unit tests for AnchorEntity fail with a ClassCastException because the test double FakeRuntimeAnchor doesn't implement the required ExportableAnchor interface. This change updates FakeRuntimeAnchor to implement the interface, adding the necessary nativePointer and anchorToken properties with dummy values. This fix unblocks unit tests for functions that require public-facing AnchorEntity by allowing AnchorEntity to be created successfully in a unit test environment.

This has to be removed once aosp/3751914 is submitted.

Relnote: Updated FakeRuntimeAnchor to implement ExportableAnchor, fixing AnchorEntity creation in unit tests.
Bug: 448467991
Test: New unit tests are added for the new behaviors.
Change-Id: I1dfaf5da53450660cdbf17e8f910bae46eb2954e