Remove @Composable from calculateScene
Since calculateScene can be called multiple
times in a loop and does not have any
defined lifetime, using any @Composable
method (e.g., rememberSaveable,
collectAsState(), etc.) is not well defined.

It may converge eventually, but is not
intuitive or recommended. Instead, any
composable work should be done at a global
level as part of the construction of the
SceneStrategy in a rememberMySceneStrategy()
method for instance.

Relnote: "The `calculateScene` method on
`SceneStrategy` is no longer `@Composable`. Instead,
consider moving this work to the construction of your
`SceneStrategy` method (i.e., in a
`rememberMySceneStrategy()` method) that fully allows
you to define the lifetime of any `rememberSaveable`
values or key `remember` statements off of the correct
values."
Fixes: 448709506
Test: existing tests and use cases still work

Change-Id: If17337d860c1d3a8b048639f2c929e8d4f9a2b10