Clean up file more reliably
Currently, the clean-up happens during try-finally block, but this isn't
reliable as if the app itself crashes, the finally block does not get
executed.

Instead, the clean-up will happen before the client callback. This way,
the clean-up will be done before the callback and any unexpected client
behaviors will not affect the clean-up.

Test: Tested by crashing test app and observing no residuals
Change-Id: Ie41477db5f29b102eebfa282ba278ba2f3f708e0