Add BrowserInput
Add an Input that connects browser's back and forward buttons to Navigation Event in Kotlin/Wasm.

Clicking browser's back or forward button triggers a back or forward navigation event. Clicking the dropdown list from long clicking the browser back or forward button trigger one or more navigation events. For example, we have browser history:

[a, b, c*]

And then user chooses to go back to `a` using browser's UI:

[a*, b, c]

We then trigger two back navigation events.

The browser's history is always populated from NavigationEventInfo, and we put indices as the state object, as there's no way to know the current index using the History API. Whenever NavigationEventInfo changes, we update browser's history. Since there's no way to remove entries without pushing new entries with the History API, we treat extra entries as invalid. For example, we have NavigationEventHistory:

[a, b, c*]

We populate the browser history as (# denotes the end of valid entries):

[0, 1, 2*#]

And when we get new NavigationEventHistory [d*], we update the browser history as:

[0*#, 1, 2]

Where 0 is a valid entry, while 1 and 2 are left as-is as invalid entries. Note that we may be able to use `history.pushState` to reduce the number of invalid entries to one (so it becomes [0*#, 1]) but calling pushState may make all entries we added skippable (https://chromium.googlesource.com/chromium/src/+/refs/heads/lkgr/docs/history_manipulation_intervention.md).

When user goes to an invalid entry we directly move them back to where they came from with `history.go`.

Test: BrowserInputTest
Change-Id: I464df9c20c4f29d774d3a056cceb5ae6873f8380