YTXR needs a means of injecting the rootSceneNode and taskWindowLeashNode since it's a chimeric app. This is needed to unblock the g3 drops from AOSP.
Added a create method to SpatialSceneRuntime.java that they can call into. Marked SpatialSceneRuntime as public and restricted all other public methods.

Removed SubspaceNodeHolder and SubspaceNodeFeature along with implementations. We can defer all SubspaceNodeEntity logic to the scene runtime.

We can also handle SpatialEnvironment logic for when the SpatialEnvironmentPreference is null or contains nulls for both the ExrImage and GltfEntity inside of the SpatialEnvironmentImpl to avoid invoking logic through the SpatialEnvironmentFeature.

Tested=presubmits
Relnote=N/A
Fixes=448188531

Change-Id: Ieaecba8f2a776df686ff2c291db9f5ad7a4e9eb7