Removing currentCompositeKeyHashCode in order to lower the Compile SDK from 35 -> 34
currentCompositeKeyHashCode was not introduced until androidx.compose.runtime:runtime:1.9 so we were not able to update until we also updated the compile sdk = 35. However, we've received feedback from SS that their build system does not work with 35, so by removing it we can revert back to compile SDK 34.

We also deemed that it is not necessary to have unique identifiers for Entities as we already have a .tag() subspaceModifier for composables and it is not possible for a user to reverse engineer the tag information to track an entity.

Next, with further research and communication with @hiichoudhary@google.com he informed me if we need to use compile sdk = 34 it's best not to pass the standard test dispatcher in the test rules, because the internal test APIs need changes to support it. Therefore, I had to suppress the standard test dispatcher warning at the module level and revert the
changes made to the xr-compose module in aosp/3760708 and aosp/3766326

Test: Passes all presubmits

Bug: 450009252

Relnote: Lower jxr-compose modules to Compile sdk = 34

Change-Id: I2d5db3bfc2acd903fbd76d7fb677170aa3de84ee