Prevent crash when destroying an Activity with a Subspace
When an Activity containing an ApplicationSubspace was destroyed, it could lead to a race condition during teardown when `ApplicationSubspace` attempts to re-enable the `mainPanelEntity` after the `SpatialComposeScene` is disposed. However, if the Activity is already in the `DESTROYED` state, the underlying XR `Session` is also destroyed.

This change resolves the crash by wrapping a `try-catch` block. This safely handles the `IllegalStateException` that occurs when the session is destroyed concurrently with the composition's disposal. `context.requiredActivity().isFinishing` isn't helping in this case because of the race condition, where the Lifecycle.State can be updated to DESTROYED moments before the Activity's internal isDestroyed or isFinishing flags are updated.

The catch block is intentionally left empty, as re-enabling the main panel is unnecessary if the entire application is being torn down. A comment has been added to explain why this exception can be safely ignored.

Bug: 449218313
Test: Manually verified in Unified test app and Animation app in emulator.
Change-Id: I837670c52f69f43ad9e95059d55da36251054b2d