Fix concurrency bug in GuavaDataStore.
This change fixes an issue where `GuavaDataStore` operations could incorrectly run on the calling thread (e.g., the main thread) instead of the specified IO dispatcher. The root cause was `launchFuture` in GuavaDataStore starting coroutines with `CoroutineStart.UNDISPATCHED` by default, which led `withContext` in the `readState` of `DataStoreImpl` to take a fast-path optimization and skip the necessary thread switch from the Main thread to the I/O thread. The fast-path optimization was caused by the fact that the current `CoroutineContext` already contained the `IO` dispatcher, and therefore the switch from the Main thread to the IO thread was skipped.

By explicitly setting `launchUndispatched = false`, we ensure the coroutine is always scheduled on the correct background dispatcher, preventing blocking I/O on the main thread. The test `GuavaDataStoreDispatchTest` asserts that `getDataAsync()` and `updateDataAsync()` are not called on the Main thread and we are switched to the IO thread as expected.

Bug: b/441801112
Test: GuavaDataStoreDispatchTest.kt
Change-Id: Ic91eab1f61907b2c3d97bca18b43ce7c9d7658a3