Remove room.generateKotlin processor option.
This change removes the ability to generate Java code in Room, including removing the processor option from room-compiler and room-gradle-plugin. It also removes Room's JavacBasicAnnotationProcessor, meaning only the a KSP version of Room's compiler will remain generating Kotlin.

Due to KSP having improved nullability various tests had to be updated to reflect error messages that reference input types that are nullable. Moreover since the Java AP processor is removed, tests that relied on had to be refactored to use the KSP version.

Bug: 438041176
Test: Existing
Change-Id: I9754e1b9595a926dc2d6023dc77bcfb407e09bad