Migrate Room's AutoValue test app to KSP
The test app validates the AutoValue support in Room, however, AutoValue is a Java Annotation Processor that does not support KSP and it cannot interact in multi-round fashion with Room KSP. Therefore instead of losing the coverage all of the AutoValue generated sources are checked-in as sources of the test app. Since the dependency between AutoValue and Room KSP is simple, this works to continue validating the support because there are *some* custom infra out there that does have KAPT and KSP interop.

Bug: 438041176
Test: ./gradlew :room3:integration-tests:room-testapp-autovalue:cAT
Change-Id: I84060c0c609931e45c1c026e106fcf076d4d2f6e