Code refactor: Introduce CameraGraphConfigProvider
This CL refactors the creation of CameraGraph.Config
out of UseCaseManager and into a new, dedicated
CameraGraphConfigProvider.

The UseCaseManager class was responsible for both
managing UseCase lifecycles and the complex, low-level
logic of building a CameraGraph.Config. This was done
via a large static utility method with a long parameter
list, which cluttered UseCaseManager's responsibilities
and its constructor with dependencies that were only
used for this creation task.

A new, camera-scoped CameraGraphConfigProvider is introduced.
This class now encapsulates all the logic and dependencies
required to build a valid CameraGraph.Config. UseCaseManager
now injects this provider and delegates the configuration
creation to it.

Bug: 448593362
Test: UseCaseManagerTest
Change-Id: I370bc2e23e008348abcfbdbd6b299005fc01560b