Add compiler generated versions of Kotlin functions with optional parameters
When a Kotlin function has parameters with default values, the compiler generates a version which ends up used when the source function is called without all parameters from Kotlin source.

Previously, metalava was not tracking the compiler generated default version, but it should be since it would be binary breaking if the signature were to change. For more details of the issue see this document: https://docs.google.com/document/d/1E1gYBNsgEUyduiEQ7yMd0o1DQ224vCNVtwqxPfBts_I/edit

Bug: 432049122
Test: ./gradlew checkApi
Relnote: N/A
Change-Id: I661eac2a35b3b0de0cb6d13d1f05e068e4b1deb7