[2Page] Introduce a layout strategy interface.
This CL introduces a `LayoutStrategy` interface to define how pages are positioned and which pages are visible within the viewport. This abstraction will allow for the implementation of different layout managers, such as a single-page view or a two-page spread.

This change is a foundational step towards supporting more complex layouts in the PDF viewer.

Bug: 449726514
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I169524ba5f4d911ba8a140f25fd59a4d3f79b436