Migrate pre-packaged database feature to driver APIs.
This change migrates Room's pre-package database functionality (i.e. the copyFrom* APIs from the database builder) to support a Room database that has a driver installed. The SupportSQLite version is then removed. This change does not make the API KMP, but enables driver users to continue using the functionality.

Bug: 339934813
Bug: 438041176
Test: PrepackageTest.kt
Relnote: Add an overload in PrepackagedDatabaseCallback that receives a SQLiteConnection to support invoking the callback when a driver is configured in the RoomDatabase.
Change-Id: I26d361ae413d25a6db5bf2191637f92d10823cf6