Create account store in AppSearch
This change add a AppSearchAccount as a super class of GenericDocument,
this class will be used in the framework when the built-in type is not
available.

The account store is the place to store accounts that has reference
documents in AppSearch. The account property paths is stored in icing
schema log.

DD: go/appsearch-account-storage
Bug: 413089233
Test: AccountStoreTest

Change-Id: I80d9f810e1fe4588fba571f1a3c50dbd145d9b72