Fix for GraphicsLayerOwnerLayer to not reset to identity matrix state
There was a mistake of providing 1f instead of 0f as a default value for translationZ, resulted in layers to never getting back to isIdentity==true state after any matrix transformation. I didn't come up with a way to write a proper test as we don't expose this is identity property anywhere in the public api, and the behavior was still correct, it was only a performance issue, as we were applying matrix transformations where they might be avoided.

Test: manually via debugger
Change-Id: I13765c8e968c0ed2d82490abc4725c400ea13334