Reschedule periodic work that throws an exception.
Previously, periodic workers that threw an uncaught exception would be
marked as FAILED and not be rescheduled. This change ensures that when
a periodic worker fails with an exception, it is correctly reset and
rescheduled for its next run, similar to how `Result.failure()` is
handled.

Relnote: "Fixed an issue where periodic `WorkRequest`s that fail due to an uncaught exception would not be rescheduled."
Bug: 443879071
Test: ./development/validate_changes.sh HEAD~1 buildOnServer,test,connectedAndroidTest --runOnDependentProjects
Change-Id: Ia6b9216b669b67441f0d429112059faf8fa806c9