Update Room Incremental AP Test for KSP 2.0.4
Specifically use the newer Gradle plugin syntax which requires using GradleTestKit's plugin classpath discovery. Moreover the Kotlin classes output directory seem to have changed and because the test validates which files are changed on compilation tasks the location path is updated. Lastly JVM target is set to 11 to avoid the 1.8 deprecation warning.

Bug: 438508447
Test: RoomIncrementalAnnotationProcessingTest

Change-Id: I43328d6195e56ab8fe41723f7565b9851ed01d8e