New LocalAmbientModeManager composition
LocalAmbientModeManager is the primary way for any composable in the hierarchy to access the manager responsible for tracking the ambient (low-power) mode state.
Composables can use the new local composition to monitor device ambient state and subscribe to UI updates at infrequent intervals when device is in ambient mode.

Bug: 427724331
Test: Samples and manual verification
Relnote: Introduced LocalAmbientModeManager, a new composable to easily define different UI and behavior for interactive and ambient modes.

Change-Id: I0016170b1ef44293d526654ed8aeadd4204415ae