Make CorruptionHandler public.
As per API Council guidelines, internal super type for public types are not allowed. The CorruptionHandler interface has been made public to abide by this rule as it cannot be made a `sealed interface` due to the the fact that it has KMP expect/actuals to support the presence of the `@Throws` annotation on the handleCorruption() function in the jvm targets.

Deleting CorruptionHandler is not an option as it is needed as a super type for NoOpCorruptionHandler to be used "under the hood" when a ReplaceFileCorruptionHandler is not provided.

Bug: 452406457
Relnote: Make CorruptionHandler public.
Test: Should not break existing usages, must work with updateApi and updateAbi.
Change-Id: I9ac35e246746353251965c34b8de3bb2a5ea09aa