Fix screen flash when swapping backStacks
When swapping backStacks in NavDisplay, the NavEntry occupying the old position in compose has been replaced by the new NavEntry in the new backstack. Therefore when compose recomposes the old NavEntry as it animates out, it actually invokes the content of the new NavEntry instead.

Now we key the NavEntry instantiation within `decorateEntry` block to help compose differentiate between the different entries. We could have keyed this blocked within `rememberDecoratedNavEntries` where it actually loops through all entries to decorate them, but in the case where users swap out decorators as well to preserve the previous stack's states, the old entries' DisposableEffect needs to be aware that decorators have been swapped out (it needs access to updated decorators). This means the updated decorators need to be outside of this key block in order for the old entry to gain access. So instead, we key the actual instantiation of NavEntry.

Bug: 450967248
Test: ./gradlew navigation3:navigation3-ui:androidConnectedCheck
Relnote: "Fixed screen flicker when swapping the backStack passed to NavDisplay"
Change-Id: Ief7b5f0a334f15f51fd6ee0d81b0af39c46198f2