Refine SessionConfig and LifecycleCameraProvider docs
Improves the discoverability of key constraints in `SessionConfig`
by moving detailed explanations for `useCases` and `frameRateRange`
from property-specific tags to the main class-level KDoc. This
ensures better visibility in generated documentation, where property
comments can be less prominent. For example, see
https://developer.android.com/reference/androidx/camera/core/SessionConfig#getFrameRateRange

Additionally, this change clarifies the behavior of `isBound` and
`unbind` in `LifecycleCameraProvider` and `ProcessCameraProvider`,
specifying that they operate on exact `SessionConfig` instances.

Bug: 452536150
Test: Manual verification of generated documentation.
Change-Id: I3051ef2262211cf1de267a3c9f29076e920ca0f9