Manual overload instead of default parameter in open @Composable fun
This change avoids using default parameter in rememberSharedContentState
composable function by creating manual overloads. This is needed
because kotlin 2.1 has an issue where overriding open
composable functions would lead to a crash. This is fixed in
kotlin 2.2. But we are defensively changing the API so it's
functional for devs using kotlin 2.1.

RelNote: N/A
Fixes: 452633515
Test: ./gradlew :compose:animation:animation:updateApi
Change-Id: I4692b382a760b79eac57b9ec075cf23cf1f71a6b