Share configuration cache between postsubmit and presubmit
To achieve this, `org.gradle.configuration-cache.entries-per-key` is set to 2. This allows Gradle to store separate cache entries for the two values of the IS_POSTSUBMIT env var.

When storing the cache entry for presubmit, we run gradle with `--dry-run` as we're only interesting in storing the config cache entry, and not in executing the tasks

TESTED:
When I run the following commands im storing the entries to the same key
IS_POSTSUBMIT=true ./gradlew :tasks --dry-run
IS_POSTSUBMIT=false ./gradlew :tasks --dry-run

When I run the commands again, the config cache is no longer invalidated due to IS_POSTSUBMIT changing

Change-Id: I20eadd2483ff248fc71e33885a2a14d8d5182a47