Fix incorrect parameters passed to FeatureSettings
Fixes a bug in UseCaseManager where parameters were passed in the
wrong order to the `SupportedSurfaceCombination.FeatureSettings`
constructor. The `isPreviewStabilizationOn` value was incorrectly
passed for the `hasVideoCapture` parameter, and `isUltraHdrOn` was
passed for `isPreviewStabilizationOn`.

This change corrects the order and adopts named arguments to prevent
future errors.

As part of this fix, the `isVideoCapture` and `containsVideoCapture`
utility methods were moved to a new `UseCaseUtil.kt` file to improve
code organization.

Bug: 453931373
Test: Manual check and running all local host tests
Change-Id: I140a4e432a4aed3b47ec5331902994c816e58223