File handling: grant file permissions to browser based on a check
Currently we assume that the app has both read and write file permissions.
If the app has read-only permissions, granting read and write
permissions to the browser will fail.
However, we should also allow opening read-only file in a web app and
letting the developers handle write permission check on the app side.

Test: No additional tests; relevant tests don't require update
Bug: 416438475
Change-Id: I827cc81694f783a85f6b7ee214fe58a4e02c9a62