Format Floating-Point Data in Integration Test App
The panel in the integration test app displays floating-point data like distances and dimensions. Showing the full-precision values is meaningless and hard to read at a glance, which causes QA to easily misunderstand the data and report non-existent bugs. This patch formats the data to two decimal places to prevent further confusion.

Bug: b/416114988
Test: On the Moohan device
Change-Id: I43030b5c1bf6ad92d50bb6367a82e3668e7d67d2