Update AndroidX signature files
There were two Metalava changes associated with this signature files update.

The first was adding a suppress compatibility annotation to file facade
classes if all the members of a file facade class are marked as
experimental.

The second was adding a suppress compatibility annotation to packages in
signature files if all the top-level classes in the API surface are
marked as experimental.

Bug: 404795417 and 408977387
Test: N/A
Relnote: N/A
Change-Id: I109a68358dd0337185910d287bd55c887ac4a494