[glance] import canonical layouts to integration tests
This ports the canonical layouts into the glance integration test
folder. The port was done with the help of gemini cli and is intended
to serve as a visual integration test for the remote compose work.
These demos should render the same using either the RV or RC backends.

Lint errors have been suppressed. This is a temporary error that will
be resolved when the feature passes api review and the apis are made
public.

Test: manually build and run.

Change-Id: Ic47c641e6f216a6580bf56f38685b3dd791cdb1d