Replace LiveData machinery with Flow.asLiveData()
This change replaces RoomTrackingLiveData and its container by creating an invalidation tracker flow and using the asLiveData() conversion function.

This change is done to eliminate usages of the Observer APIs as it is meant to be removed, replaced by the first party support of Flows the tracker has.

LiveData support in Room 3.0 is pending the development of custom DAO return types and it is planned to be moved to a helper artifact.

Bug: 329315924
Bug: 438041176
Test: Existing
Relnote: "Remove the no longer used code generation helper createLiveData() that took a Callable."
Change-Id: I61af9b576e756718465a537e975be243aba0b6e0