Fix swiping in SwipeDismissableNavHostBenchmark
In order to dismiss the screen, user has to scroll for
at least 50% of the screen width. Benchmark tried to do
exactly that, but the problem is that, depending on
steps and screen resolution, swiping distance was not
always exactly 50%. In some cases it was lower.
Change swiping distance to 75% of screen width to go
around corner cases

Bug: 455008184
Test: SwipeDismissableNavHostBenchmark passes
Relnote: N/A

Change-Id: I919e772714f335883af68c00a0c0b830b19b7c08